pyproject.toml
[project]
name = "iam-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.5",
    "httpx>=0.27.2",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.32.0",
]

[dependency-groups]
dev = [
    "ruff>=0.7.4",
]
